Switch to raw input mode during interactive REPL prompt
This will allow us to intercept input events to perform actions once
triggered. For now, we just handle the existing capabilities of the
REPL, which include appending and deleting characters from the prompt.

Raw input mode is widely supported across terminals, though its
activation requires different APIs in Windows vs POSIX-compliant
systems. We opt to only enter this mode when being prompted for a
command. Once a command is being executed, we revert back to the
starting input mode that was present when the application was started.

As user inputs are being provided in this mode, we need to accurately
track the terminal's size and cursor position to handle certain complex
use-cases (i.e., inserting/deleting a character in the middle of the
prompt buffer, navigating through the command history, etc.).